+++ e2e/addressbook-tests.spec.js
@@ -4,6 +4,7 @@ import TestHelpers from './helpers';
 const INVALID_ADDRESS = '0xB8B4EE5B1b693971eB60bDa15211570df2dB221L';
 const MYTH_ADDRESS = '0x1FDb169Ef12954F20A15852980e1F0C122BfC1D6';
 const MEMO = 'Test adding ENS';
+const PASSWORD = '12345678';
 
 describe('Addressbook Tests', () => {
 	beforeEach(() => {
@@ -17,8 +18,38 @@ describe('Addressbook Tests', () => {
 		await TestHelpers.waitAndTap('onboarding-get-started-button');
 		// Check that we are on the onboarding screen
 		await TestHelpers.checkIfVisible('onboarding-screen');
-		// Check that Start Exploring CTA is visible & tap it
-		await TestHelpers.waitAndTap('start-exploring-button');
+		// Check that Create a new wallet CTA is visible & tap it
+		await TestHelpers.waitAndTap('create-wallet-button');
+		// Check that we are on the Create password screen
+		await TestHelpers.checkIfVisible('create-password-screen');
+		// Input new password
+		await TestHelpers.typeTextAndHideKeyboard('input-password', PASSWORD);
+		// Input confirm password
+		await TestHelpers.typeTextAndHideKeyboard('input-password-confirm', PASSWORD);
+		// Mark the checkbox that you understand the password cannot be recovered
+		if (device.getPlatform() === 'ios') {
+			await TestHelpers.tap('password-understand-box');
+		} else {
+			// Tap by the I understand text
+			await TestHelpers.delay(1000);
+			await TestHelpers.tap('i-understand-text');
+		}
+		// Tap on create password button
+		await TestHelpers.tap('submit-button');
+		// Check that we are on the Secure your wallet screen
+		await TestHelpers.checkIfVisible('protect-your-account-screen');
+		// Tap on the remind me later button
+		await TestHelpers.tap('remind-me-later-button');
+		// Check the box to state you understand
+		if (device.getPlatform() === 'ios') {
+			await TestHelpers.tap('skip-backup-check');
+		} else {
+			// Tap by the I understand text
+			await TestHelpers.delay(1000);
+			await TestHelpers.tap('skip-backup-text');
+		}
+		// Tap on Skip button
+		await TestHelpers.tapByText('Skip');
 		// Check that we are on the metametrics optIn screen
 		await TestHelpers.checkIfVisible('metaMetrics-OptIn');
 		// Check that I Agree CTA is visible and tap it
@@ -34,6 +65,10 @@ describe('Addressbook Tests', () => {
 		await TestHelpers.waitAndTap('onboarding-wizard-back-button');
 		// Check that the onboarding wizard is gone
 		await TestHelpers.checkIfNotVisible('onboarding-wizard-step1-view');
+		// Check that the protect your wallet modal is visible
+		await TestHelpers.checkIfVisible('backup-alert');
+		// Tap on remind me later
+		await TestHelpers.tap('notification-remind-later-button');
 	});
 
 	it('should go to send view', async () => {
@@ -143,6 +178,7 @@ describe('Addressbook Tests', () => {
 		if (device.getPlatform() === 'android') {
 			await TestHelpers.tapByText('Add contact');
 			await TestHelpers.delay(1000);
+			await TestHelpers.tapByText('Add contact');
 		} else {
 			await TestHelpers.tap('contact-add-contact-button');
 			await TestHelpers.tap('contact-add-contact-button');